refactor(demo): 重构 src 目录组织为 view/components，并合并全局样式
- 视图迁移至 src/view/*：JsonPatchDemo、JsonPatchEditor、ConflictResolutionSection、PatchPreviewSection、ResultSection、SchemaEditSection
- 保留复用组件于 src/components/*：JsonEditor、OptionSelect、DataStatusIndicator、WelcomeMessage
- 新增别名 @src 指向 src（vite.config.ts、tsconfig.json）
- 合并样式为 src/styles/app.css；移除 src/index.css 与 src/components/styles.css
- 更新入口与引用：
  - App.tsx -> 使用 @src/view/JsonPatchDemo
  - main.tsx -> 仅引入 @src/styles/app.css
- 清理废弃文件：删除旧的 view 组件在 components 中的转发文件
- 无功能逻辑变更；构建通过（pnpm -w build）

Refs: code organization, modern structure, consistency
