--- packages/demo/src/index.css
@@ -1,13 +1,0 @@
-@import "tailwindcss";
-
-.container {
-    max-width: 1200px;
-    margin: 24px auto;
-    padding: 0 16px;
-}
-
-@media (max-width: 768px) {
-    .container {
-        margin: 16px auto;
-    }
-}
