+++ packages/demo/src/main.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import './index.css';
-import './components/styles.css';
+import '@src/styles/app.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
