+++ packages/demo/src/styles/app.css
@@ -1,3 +1,18 @@
+@import "tailwindcss";
+
+/* 布局容器 */
+.container {
+    max-width: 1200px;
+    margin: 24px auto;
+    padding: 0 16px;
+}
+
+@media (max-width: 768px) {
+    .container {
+        margin: 16px auto;
+    }
+}
+
 /* 通用样式 */
 .json-patch-demo {
     max-width: 1200px;
@@ -206,3 +221,5 @@
 .json-editor textarea {
     font-family: 'Courier New', Courier, monospace;
 }
+
+
