+++ packages/demo/src/view/ResultSection.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Card, Button, Divider, Space, Typography } from 'antd';
 import { ReloadOutlined } from '@ant-design/icons';
-import JsonEditor from './JsonEditor';
+import JsonEditor from '../components/JsonEditor';
 import { usePatchContext } from '../context/PatchContext';
 
 const { Text } = Typography;
@@ -22,12 +22,7 @@ const ResultSection: React.FC = () => {
 
     return (
         <div className="result-section">
-            <Card
-                title="补丁应用结果"
-                className="result-card"
-                bordered={false}
-                style={{ borderRadius: '8px', boxShadow: '0 2px 8px rgba(0, 0, 0, 0.08)' }}
-            >
+            <Card title="补丁应用结果" className="result-card" bordered={false} style={{ borderRadius: '8px', boxShadow: '0 2px 8px rgba(0, 0, 0, 0.08)' }}>
                 <Typography>
                     <Text>以下是应用补丁后的结果。您可以比较源JSON和结果JSON之间的差异。</Text>
                 </Typography>
@@ -35,35 +30,15 @@ const ResultSection: React.FC = () => {
                 <Divider orientation="left" style={{ margin: '24px 0 16px' }}>
                     源数据
                 </Divider>
-                <JsonEditor
-                    value={sourceJson}
-                    onChange={() => {}}
-                    readOnly={true}
-                    height="200px"
-                    title="源JSON"
-                />
+                <JsonEditor value={sourceJson} onChange={() => {}} readOnly={true} height="200px" title="源JSON" />
 
                 <Divider orientation="left" style={{ margin: '24px 0 16px' }}>
                     应用补丁后的结果
                 </Divider>
-                <JsonEditor
-                    value={resultJson}
-                    onChange={() => {}}
-                    readOnly={true}
-                    height="200px"
-                    title="结果JSON"
-                />
+                <JsonEditor value={resultJson} onChange={() => {}} readOnly={true} height="200px" title="结果JSON" />
 
                 {error && (
-                    <div
-                        className="error-message"
-                        style={{
-                            marginTop: '16px',
-                            padding: '8px 12px',
-                            backgroundColor: '#fff2f0',
-                            borderRadius: '4px',
-                        }}
-                    >
+                    <div className="error-message" style={{ marginTop: '16px', padding: '8px 12px', backgroundColor: '#fff2f0', borderRadius: '4px' }}>
                         <Text type="danger">{error}</Text>
                     </div>
                 )}
@@ -72,12 +47,7 @@ const ResultSection: React.FC = () => {
 
                 <div className="actions">
                     <Space>
-                        <Button
-                            type="primary"
-                            onClick={resetWorkflow}
-                            icon={<ReloadOutlined />}
-                            size="middle"
-                        >
+                        <Button type="primary" onClick={resetWorkflow} icon={<ReloadOutlined />} size="middle">
                             重新开始
                         </Button>
                     </Space>
@@ -88,3 +58,5 @@ const ResultSection: React.FC = () => {
 };
 
 export default ResultSection;
+
+
