+++ packages/demo/src/view/SchemaEditSection.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Card, Button, Space, Typography } from 'antd';
 import { RightOutlined, InfoCircleOutlined } from '@ant-design/icons';
-import JsonEditor from './JsonEditor';
+import JsonEditor from '../components/JsonEditor';
 import { usePatchContext } from '../context/PatchContext';
 
 const { Text, Paragraph } = Typography;
@@ -18,20 +18,13 @@ const SchemaEditSection: React.FC = () => {
                 title="数据模型定义"
                 className="schema-card"
                 extra={
-                    <Button
-                        type="primary"
-                        onClick={() => setActiveTab('editor')}
-                        icon={<RightOutlined />}
-                        size="middle"
-                    >
+                    <Button type="primary" onClick={() => setActiveTab('editor')} icon={<RightOutlined />} size="middle">
                         下一步
                     </Button>
                 }
             >
                 <Space direction="vertical" className="schema-content" size="large">
-                    <div
-                        style={{ backgroundColor: '#f0f7ff', padding: '12px', borderRadius: '6px' }}
-                    >
+                    <div style={{ backgroundColor: '#f0f7ff', padding: '12px', borderRadius: '6px' }}>
                         <Paragraph style={{ marginBottom: '8px' }}>
                             <InfoCircleOutlined style={{ color: '#1677ff', marginRight: '8px' }} />
                             请定义您的数据模型，用于指导补丁的生成和应用。
@@ -42,22 +35,10 @@ const SchemaEditSection: React.FC = () => {
                         </Paragraph>
                     </div>
 
-                    <JsonEditor
-                        value={schemaString}
-                        onChange={updateSchema}
-                        title="Schema JSON"
-                        height="400px"
-                        showSaveIndicator={true}
-                    />
+                    <JsonEditor value={schemaString} onChange={updateSchema} title="Schema JSON" height="400px" showSaveIndicator={true} />
 
                     {error && error.includes('Schema') && (
-                        <div
-                            style={{
-                                padding: '8px 12px',
-                                backgroundColor: '#fff2f0',
-                                borderRadius: '4px',
-                            }}
-                        >
+                        <div style={{ padding: '8px 12px', backgroundColor: '#fff2f0', borderRadius: '4px' }}>
                             <Text type="danger">{error}</Text>
                         </div>
                     )}
@@ -68,3 +49,5 @@ const SchemaEditSection: React.FC = () => {
 };
 
 export default SchemaEditSection;
+
+
