+++ packages/demo/vite.config.ts
@@ -14,6 +14,7 @@ export default defineConfig({
     resolve: {
         alias: {
             '@': resolve(__dirname, './src'),
+            '@src': resolve(__dirname, './src'),
         },
     },
 });
